Changed text to one sentence per line.
Modified all existing text so that each sentence is on a single line. Footnotes at the end of a sentence are also on their own line, with a comment character at the end of the host sentence so that no extra whitespace is added. This should make it easier to change existing text since only one sentence rather than a whole paragraph is affected by a modification to the source code.